SLVS-1853 Move logic of Funcs into separate class
